+++ common/src/main/java/net/moritz_htk/better_mcdonalds_mod/block/BMMLettuceCropBlock.java
@@ -3,18 +3,22 @@ package net.moritz_htk.better_mcdonalds_mod.block;
 import net.minecraft.world.level.ItemLike;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.CropBlock;
+import net.minecraft.world.level.block.SoundType;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.block.state.StateDefinition;
+import net.minecraft.world.level.block.state.properties.BlockStateProperties;
 import net.minecraft.world.level.block.state.properties.IntegerProperty;
+import net.minecraft.world.level.material.MapColor;
+import net.minecraft.world.level.material.PushReaction;
 import net.moritz_htk.better_mcdonalds_mod.registry.BMMItems;
 import org.jetbrains.annotations.NotNull;
 
 public class BMMLettuceCropBlock extends CropBlock {
     public static final int MAX_AGE = 3;
-    public static final IntegerProperty AGE = IntegerProperty.create("age", 0, 3);
+    public static final IntegerProperty AGE = BlockStateProperties.AGE_3;
 
     public BMMLettuceCropBlock(Properties properties) {
-        super(properties.noCollission().noOcclusion());
+        super(properties.mapColor(MapColor.PLANT).noCollission().randomTicks().instabreak().sound(SoundType.CROP).pushReaction(PushReaction.DESTROY).noOcclusion());
     }
 
     @Override
@@ -23,7 +27,7 @@ public class BMMLettuceCropBlock extends CropBlock {
     }
 
     @Override
-    public @NotNull IntegerProperty getAgeProperty() {
+    protected @NotNull IntegerProperty getAgeProperty() {
         return AGE;
     }