+++ common/src/main/java/net/moritz_htk/better_mcdonalds_mod/registry/BMMBlocks.java
@@ -4,6 +4,8 @@ import com.mojang.serialization.MapCodec;
 import dev.architectury.registry.registries.DeferredRegister;
 import dev.architectury.registry.registries.RegistrySupplier;
 import net.minecraft.core.registries.Registries;
+import net.minecraft.resources.ResourceKey;
+import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.item.DyeColor;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.Blocks;
@@ -16,12 +18,16 @@ import net.moritz_htk.better_mcdonalds_mod.block.BMMTomatoCropBlock;
 public class BMMBlocks {
     public static final DeferredRegister<Block> BLOCKS = DeferredRegister.create(BetterMcDonaldsMod.MOD_ID, Registries.BLOCK);
 
-    public static final RegistrySupplier<FallingBlock> SALT_BLOCK = BLOCKS.register("salt_block", () -> new FallingBlock(BlockBehaviour.Properties.ofFullCopy(Blocks.SAND).mapColor(DyeColor.WHITE)) {
+    public static final RegistrySupplier<FallingBlock> SALT_BLOCK = BLOCKS.register("salt_block", () -> new FallingBlock(BlockBehaviour.Properties.ofFullCopy(Blocks.SAND).mapColor(DyeColor.WHITE).setId(registerBlockId("salt_block"))) {
         @Override
         protected MapCodec<? extends FallingBlock> codec() {
             return null;
         }
     });
-    public static final RegistrySupplier<BMMTomatoCropBlock> TOMATO_CROP = BLOCKS.register("tomato_crop", () -> new BMMTomatoCropBlock(BlockBehaviour.Properties.ofFullCopy(Blocks.WHEAT)));
-    public static final RegistrySupplier<BMMLettuceCropBlock> LETTUCE_CROP  = BLOCKS.register("lettuce_crop", () -> new BMMLettuceCropBlock(BlockBehaviour.Properties.ofFullCopy(Blocks.WHEAT)));
+    public static final RegistrySupplier<BMMTomatoCropBlock> TOMATO_CROP = BLOCKS.register("tomato_crop", () -> new BMMTomatoCropBlock(BlockBehaviour.Properties.of().setId(registerBlockId("tomato_crop"))));
+    public static final RegistrySupplier<BMMLettuceCropBlock> LETTUCE_CROP  = BLOCKS.register("lettuce_crop", () -> new BMMLettuceCropBlock(BlockBehaviour.Properties.of().setId(registerBlockId("lettuce_crop"))));
+
+    public static ResourceKey<Block> registerBlockId(String id) {
+        return ResourceKey.create(Registries.BLOCK, ResourceLocation.fromNamespaceAndPath(BetterMcDonaldsMod.MOD_ID, id));
+    }
 }