+++ neoforge/src/main/java/net/moritz_htk/better_mcdonalds_mod/datagen/provider/BMMDataMapProvider.java
@@ -6,6 +6,7 @@ import net.moritz_htk.better_mcdonalds_mod.registry.BMMItems;
 import net.neoforged.neoforge.common.data.DataMapProvider;
 import net.neoforged.neoforge.registries.datamaps.builtin.Compostable;
 import net.neoforged.neoforge.registries.datamaps.builtin.NeoForgeDataMaps;
+import org.jetbrains.annotations.NotNull;
 
 import java.util.concurrent.CompletableFuture;
 
@@ -15,7 +16,7 @@ public class BMMDataMapProvider extends DataMapProvider {
     }
 
     @Override
-    protected void gather() {
+    protected void gather(HolderLookup.@NotNull Provider arg) {
         builder(NeoForgeDataMaps.COMPOSTABLES).add(BMMItems.TOMATO.getId(), new Compostable(0.65f), false).add(BMMItems.TOMATO_SEEDS.getId(), new Compostable(0.2f), false).add(BMMItems.LETTUCE.getId(), new Compostable(0.65f), false).add(BMMItems.LETTUCE_SEEDS.getId(), new Compostable(0.2f), false).build();
     }
 }