+++ changelog.md
@@ -1,2 +1,3 @@
 **Changes:**
-- Added Chocolate McFlurry
+- Updated to Minecraft 1.21.3
+- Added Tomato, Tomato Seeds, Lettuce and Lettuce Seeds to `villager_picks_up` tag